refactor(workspace): split arm ledger tests from implementation
Move the ledger unit suite into its own test module so the production file remains below the hard file-size wall, and repair the replay field references for private-item rustdoc.

(cherry picked from commit c6676bdb376fdb6c1ea90ba70a96bcb9aacc2bad)
